requirements.txt
pandas
numpy
mysql-connector-python
matplotlib
seaborn
openpyxl
jupyter